+++ Pandora/Paradigm/Primary/Auxiliary.hs
@@ -1,0 +1,5 @@
+module Pandora.Paradigm.Primary.Auxiliary where
+
+data Vertical a = Up a | Down a
+
+data Horizontal a = Leftward a | Rightward a